requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altgraph==0.17.4
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
bottle==0.13.3
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2025.4
cron-descriptor==1.4.5
cryptography==42.0.8
Cython==3.0.12
defusedxml==0.7.1
diff-match-patch==20241021
Django==4.2.16
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-filter==23.3
django-formtools==2.5.1
django-import-export==3.3.0
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-rename-app==0.1.7
django-sslserver==0.22
django-tables2==2.7.0
django-timezone-field==7.1
django-widget-tweaks==1.5.0
dnspython==2.7.0
et_xmlfile==2.0.0
Flask==3.1.0
frozenlist==1.6.0
geographiclib==2.0
geoip2==5.0.1
geopy==2.4.1
gunicorn==23.0.0
idna==3.10
imaplib2==3.6
impacket==0.12.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.3
ldap3==2.9.1
ldapdomaindump==0.10.0
lxml==5.3.2
MarkupPy==1.18
MarkupSafe==3.0.2
maxminddb==2.6.3
multidict==6.4.3
numpy==2.2.5
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.2
pandas==2.2.3
phonenumbers==9.0.4
Pillow==10.0.1
prompt_toolkit==3.0.51
propcache==0.3.1
proxy_tools==0.1.0
psycopg2-binary==2.9.7
pyarmor==7.7.4
pyarmor.cli.core==7.6.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyOpenSSL==24.0.0
pypng==0.20220715.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.16
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==0.8.11
python-dotenv==1.1.0
pytz==2025.2
pywebview==5.4
PyYAML==6.0.2
qrcode==7.4.2
redis==6.0.0
reportlab==4.0.4
requests==2.32.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tablib==3.5.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.9.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.20.0
